sql: renames and moves
Rename executor.go to exec_util.go and move the contents of session.go
into it. These are now remains of code grown around the old
Session/Executor that's still used.

Release note: None
